+++ scripts/macros/classFeatures/bard/collegeOfCreation/animatingPerformance.js
@@ -91,6 +91,13 @@ async function item({speaker, actor, token, character, item, args, scope, workfl
             }
         }
     };
+    let avatarImg = chris.getConfiguration(workflow.item, 'avatar');
+    if (avatarImg) updates.actor.img = avatarImg;
+    let tokenImg = chris.getConfiguration(workflow.item, 'token');
+    if (tokenImg) {
+        setProperty(updates, 'actor.prototypeToken.texture.src', tokenImg);
+        setProperty(updates, 'token.texture.src', tokenImg);
+    }
     await tashaSummon.spawn(sourceActor, updates, 3600, workflow.item);
     let featureData = await chris.getItemFromCompendium('chris-premades.CPR Class Feature Items', 'Animating Performance - Command', false);
     featureData.system.description.value = chris.getItemDescription('CPR - Descriptions', 'Animating Performance - Command');