+++ scripts/macros/spells/summonCelestial.js
@@ -51,13 +51,12 @@ export async function summonCelestial({speaker, actor, token, character, item, a
         'embedded': {
             'Item': {
                 [multiAttackFeatureData.name]: multiAttackFeatureData,
-                [healingTouchData.name]: healingTouchData,
-                'Configure Images': warpgate.CONST.DELETE
+                [healingTouchData.name]: healingTouchData
             }
         }
     };
-    let avatarImg = sourceActor.flags['chris-premades']?.summon?.image?.[selection]?.avatar;
-    let tokenImg = sourceActor.flags['chris-premades']?.summon?.image?.[selection]?.token;
+    let avatarImg = chris.getConfiguration(workflow.item, 'avatar-' + selection);
+    let tokenImg = chris.getConfiguration(workflow.item, 'token-' + selection);
     if (avatarImg) updates.actor.img = avatarImg;
     if (tokenImg) {
         setProperty(updates, 'actor.prototypeToken.texture.src', tokenImg);