+++ scripts/macros/spells/summonShadowspawn.js
@@ -59,13 +59,12 @@ export async function summonShadowspawn({speaker, actor, token, character, item,
             'Item': {
                 [multiAttackFeatureData.name]: multiAttackFeatureData,
                 [chillingRendData.name]: chillingRendData,
-                [dreadfulScreamData.name]: dreadfulScreamData,
-                'Configure Images': warpgate.CONST.DELETE
+                [dreadfulScreamData.name]: dreadfulScreamData
             }
         }
     };
-    let avatarImg = sourceActor.flags['chris-premades']?.summon?.image?.[selection]?.avatar;
-    let tokenImg = sourceActor.flags['chris-premades']?.summon?.image?.[selection]?.token;
+    let avatarImg = chris.getConfiguration(workflow.item, 'avatar-' + selection);
+    let tokenImg = chris.getConfiguration(workflow.item, 'token-' + selection);
     if (avatarImg) updates.actor.img = avatarImg;
     if (tokenImg) {
         setProperty(updates, 'actor.prototypeToken.texture.src', tokenImg);