+++ scripts/module.js
@@ -93,7 +93,6 @@ Hooks.once('ready', async function() {
         if (game.modules.get('itemacro')?.active) {
             if (game.settings.get('itemacro', 'charsheet')) ui.notifications.error('Chris\'s Premades & Midi-Qol requires "Character Sheet Hook" in Item Macro\'s module settings to be turned off!');
         }
-        Hooks.on('getItemSheetHeaderButtons', createHeaderButton);
         if (game.modules.get('ddb-importer')?.active) Hooks.on('getActorSheet5eHeaderButtons', createActorHeaderButton);
         game.settings.set('chris-premades', 'LastGM', game.user.id);
         if (game.settings.get('chris-premades', 'Combat Listener')) Hooks.on('updateCombat', combatUpdate);
@@ -157,6 +156,9 @@ Hooks.once('ready', async function() {
         Hooks.on('midi-qol.preCheckHits', macros.manualRolls.attackRoll);
         Hooks.on('midi-qol.postCheckSaves', macros.manualRolls.saveRolls);
         await patching();
+    }
+    if (game.user.isGM || game.settings.get('chris-premades', 'Item Replacer Access') || game.settings.get('chris-premades', 'Item Configuration Access')) {
+        Hooks.on('getItemSheetHeaderButtons', createHeaderButton);
     }
 });
 globalThis['chrisPremades'] = {